Улучшена система бонусов, больше не рандом, меняются в зависимости от суммы оплаты. Добавлен фрагмент оплаты с эмуляцией отказа

+++ app/src/main/java/com/example/coffeefirst/data/model/MenuItem.kt
@@ -5,5 +5,5 @@ data class MenuItem(
     val name: String,
     val imageResId: Int,
     val category: String,
-    val price: Double
+    val price: Float
 )
